pyproject.toml
[project]
name = "multiagent"
version = "0.1.0"
description = "LLM Multi-Agent Application Framework"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    # LLM & Agent Frameworks
    "anthropic>=0.39.0",
    "langgraph>=0.2.0",
    "langchain>=0.3.0",
    "langchain-anthropic>=0.3.0",
    "langchain-community>=0.3.0",
    "google-generativeai>=0.8.0",

    # API Framework
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.6.0",

    # Vector Database
    "chromadb>=0.5.0",

    # State & Caching
    "redis>=5.2.0",

    # Azure Storage
    "azure-storage-blob>=12.23.0",
    "azure-identity>=1.19.0",

    # Async Support
    "httpx>=0.27.0",
    "aiofiles>=24.1.0",

    # Utilities
    "python-dotenv>=1.0.0",
    "structlog>=24.4.0",
    "tenacity>=9.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "ruff>=0.7.0",
    "mypy>=1.13.0",
    "pre-commit>=4.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
